chore(user): 移除岗位

+++ wzkris-modules/wzkris-modules-user/wzkris-user-biz/src/main/java/com/wzkris/user/service/impl/SysUserServiceImpl.java
@@ -4,10 +4,8 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.wzkris.common.orm.utils.DynamicTenantUtil;
 import com.wzkris.common.security.oauth2.service.PasswordEncoderDelegate;
 import com.wzkris.user.domain.SysUser;
-import com.wzkris.user.domain.SysUserPost;
 import com.wzkris.user.domain.SysUserRole;
 import com.wzkris.user.mapper.SysUserMapper;
-import com.wzkris.user.mapper.SysUserPostMapper;
 import com.wzkris.user.mapper.SysUserRoleMapper;
 import com.wzkris.user.service.SysUserService;
 import jakarta.annotation.Nullable;
@@ -32,39 +30,31 @@ public class SysUserServiceImpl implements SysUserService {
 
     private final SysUserRoleMapper userRoleMapper;
 
-    private final SysUserPostMapper userPostMapper;
-
     private final PasswordEncoderDelegate passwordEncoder;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean insertUser(SysUser user, List<Long> roleIds, List<Long> postIds) {
+    public boolean insertUser(SysUser user, List<Long> roleIds) {
         if (user.getPassword() != null && !passwordEncoder.isEncode(user.getPassword())) {
             user.setPassword(passwordEncoder.encode(user.getPassword()));
         }
         boolean success = userMapper.insert(user) > 0;
         if (success) {
-            // 新增用户与角色管理
             this.insertUserRole(user.getUserId(), roleIds);
-            this.insertUserPost(user.getUserId(), postIds);
         }
         return success;
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean updateUser(SysUser user, List<Long> roleIds, List<Long> postIds) {
+    public boolean updateUser(SysUser user, List<Long> roleIds) {
         if (user.getPassword() != null && !passwordEncoder.isEncode(user.getPassword())) {
             user.setPassword(passwordEncoder.encode(user.getPassword()));
         }
         boolean success = userMapper.updateById(user) > 0;
-        if (success) {
-            // 删除用户与角色关联
+        if (success && roleIds != null) {
             userRoleMapper.deleteByUserId(user.getUserId());
-            userPostMapper.deleteByUserId(user.getUserId());
-            // 新增用户与角色管理
             this.insertUserRole(user.getUserId(), roleIds);
-            this.insertUserPost(user.getUserId(), postIds);
         }
         return success;
     }
@@ -75,7 +65,6 @@ public class SysUserServiceImpl implements SysUserService {
         boolean success = userMapper.deleteByIds(userIds) > 0;
         if (success) {
             userRoleMapper.deleteByUserIds(userIds);
-            userPostMapper.deleteByUserIds(userIds);
         }
         return success;
     }
@@ -95,15 +84,6 @@ public class SysUserServiceImpl implements SysUserService {
             return userRoleMapper.insertBatch(list) > 0;
         }
         return false;
-    }
-
-    private void insertUserPost(Long userId, List<Long> postIds) {
-        if (CollectionUtils.isNotEmpty(postIds)) {
-            List<SysUserPost> list = postIds.stream()
-                    .map(postId -> new SysUserPost(userId, postId))
-                    .toList();
-            userPostMapper.insertBatch(list);
-        }
     }
 
     @Override
